refactor(auth): apply new dir structure to auth
