+++ apps/web/src/App.vue
@@ -27,8 +27,8 @@ import GNB from '@/common/modules/navigations/gnb/GNB.vue';
 import NoticePopup from '@/common/modules/popup/notice/NoticePopup.vue';
 import TopNotification from '@/common/modules/portals/TopNotification.vue';
 
-import MobileGuideModal from '@/services/auth/password/MobileGuideModal.vue';
-import { AUTH_ROUTE } from '@/services/auth/route-config';
+import MobileGuideModal from '@/services/auth/components/MobileGuideModal.vue';
+import { AUTH_ROUTE } from '@/services/auth/routes/route-config';
 
 const vm = getCurrentInstance()?.proxy as Vue;
