+++ apps/web/src/common/modules/navigations/gnb/modules/gnb-profile/GNBProfile.vue
@@ -22,7 +22,7 @@ import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
-import { AUTH_ROUTE } from '@/services/auth/route-config';
+import { AUTH_ROUTE } from '@/services/auth/routes/route-config';
 import { INFO_ROUTE } from '@/services/info/route-config';
 import { MY_PAGE_ROUTE } from '@/services/my-page/route-config';
