+++ apps/web/src/router/admin-domain-service-routes.ts
@@ -4,7 +4,7 @@ import { errorRoutes } from '@/router/error-routes';
 
 import { ADMINISTRATION_ROUTE } from '@/services/administration/route-config';
 import administrationRoute from '@/services/administration/routes';
-import authRoutes from '@/services/auth/routes';
+import authRoutes from '@/services/auth/routes/routes';
 
 const TotalDashboardPage = () => import('@/services/total-dashboard/TotalDashboardPage.vue');
