+++ apps/web/src/services/auth/authenticator/external/GOOGLE_OAUTH2/template/GOOGLE_OAUTH2.vue
