+++ apps/web/src/services/auth/authenticator/external/KB_SSO/authenticator/kb-auth.ts
@@ -6,7 +6,7 @@ import { store } from '@/store';
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import { Authenticator } from '@/services/auth/authenticator';
-import { AUTH_ROUTE } from '@/services/auth/route-config';
+import { AUTH_ROUTE } from '@/services/auth/routes/route-config';
 
 class KbAuth extends Authenticator {
     static async signIn(onSignInCallback, query) {
