+++ apps/web/src/services/auth/authenticator/external/KB_SSO/template/KB_SSO.vue
