+++ apps/web/src/services/auth/authenticator/local/authenticator/space-auth.ts
