--- apps/web/src/services/auth/authenticator/type.ts
@@ -1,4 +1,0 @@
-export interface Auth {
-    signIn: (...args: unknown[]) => Promise<void>;
-    signOut: () => Promise<void>;
-}
