+++ apps/web/src/services/auth/constants/password-constants.ts
@@ -1,0 +1,5 @@
+export const PASSWORD_STATUS = {
+    RESET: 'reset',
+    FIND: 'find',
+    INVALID: 'invalid',
+} as const;
