+++ apps/web/src/services/auth/types/password-type.ts
@@ -1,0 +1,10 @@
+import type { TranslateResult } from 'vue-i18n';
+
+interface PasswordFormValidation {
+    isIdValid?: boolean,
+    idInvalidText: TranslateResult,
+}
+
+export interface PasswordFormExpose {
+    validationState: PasswordFormValidation
+}
